4.21*: bloglist expansion, step9

+++ part4/blog_list/controllers/blogs.js
@@ -4,6 +4,14 @@ const { response } = require("express");
 const Blog = require("../models/blog");
 const User = require("../models/user");
 
+// const getTokenFrom = (request) => {
+//   const authorization = request.get("authorization");
+//   if (authorization && authorization.toLowerCase().startsWith("bearer ")) {
+//     return authorization.substring(7);
+//   }
+//   return null;
+// };
+
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
@@ -60,9 +68,38 @@ blogsRouter.post("/", async (request, response, next) => {
 });
 
 blogsRouter.delete("/:id", async (request, response, next) => {
+  const token = request.token;
+
+  if (!token) {
+    return response.status(400).json({
+      error: "Missing token",
+    });
+  }
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+
+  if (!decodedToken.id) {
+    return response.status(401).json({
+      error: "token invalid",
+    });
+  }
+
   try {
-    await Blog.findByIdAndRemove(request.params.id);
-    response.status(204).end();
+    const blog = await Blog.findById(request.params.id).populate("user", {
+      username: 1,
+      name: 1,
+    });
+    if (blog.user.username === decodedToken.username) {
+      try {
+        await Blog.findByIdAndRemove(request.params.id);
+        response.status(204).end();
+      } catch (exception) {
+        next(exception);
+      }
+    } else {
+      response.status(400).json({
+        error: "Unauthorized",
+      });
+    }
   } catch (exception) {
     next(exception);
   }
